+++ decidim-proposals/spec/system/proposal_wizard_spec.rb
@@ -25,7 +25,7 @@ describe "Proposal", type: :system do
            participatory_space: participatory_process)
   end
 
-  let!(:proposal_draft) { create(:proposal, :draft, component: component, title: proposal_title, body: proposal_body) }
+  let!(:proposal_draft) { create(:proposal, :draft, author: user,component: component, title: proposal_title, body: proposal_body) }
 
   let!(:compare_proposal_path) do
     Decidim::EngineRouter.main_proxy(component).compare_proposal_path(proposal_draft)
@@ -33,6 +33,10 @@ describe "Proposal", type: :system do
 
   let!(:preview_proposal_path) do
     Decidim::EngineRouter.main_proxy(component).preview_proposal_path(proposal_draft)
+  end
+
+  let!(:edit_draft_proposal_path) do
+    Decidim::EngineRouter.main_proxy(component).edit_draft_proposal_path(proposal_draft)
   end
 
   context "when creating a new proposal" do
@@ -125,6 +129,35 @@ describe "Proposal", type: :system do
 
       it "shows a modify proposal link" do
         expect(page).to have_selector("a", text: "Modify the proposal")
+      end
+    end
+
+    context "when editing a proposal draft" do
+      context "when in step_1: edit proposal draft" do
+        before do
+          visit edit_draft_proposal_path
+        end
+
+        it "show current step_1 highlighted" do
+          puts "#{edit_draft_proposal_path}"
+          within ".wizard__steps" do
+            expect(page).to have_css(".step--active", count: 1)
+            expect(page).to have_css(".step--past", count: 0)
+            expect(page).to have_css(".step--active.step_1")
+          end
+        end
+
+        it "renders a Discard button" do
+          within ".card__content" do
+            expect(page).to have_content("Discard this draft")
+          end
+        end
+
+        it "renders a Preview button" do
+          within ".card__content" do
+            expect(page).to have_content("Preview")
+          end
+        end
       end
     end
   end